🚚 Move loaders back to core
