+++ lamindb/core/__init__.py
@@ -29,7 +29,7 @@ from lamin_utils import logger
 from lamin_utils._inspect import InspectResult
 
 from .. import errors as exceptions
-from . import datasets, fields, types
+from . import datasets, types
 from ._context import Context
 from ._mapped_collection import MappedCollection
