--- lamindb/core/fields.py
@@ -1,12 +1,0 @@
-"""Fields.
-
-The field accessor of a :class:`~lamindb.core.Record`:
-
-.. autosummary::
-   :toctree: .
-
-   FieldAttr
-
-"""
-
-from lamindb.base.types import FieldAttr  # noqa: F401
